backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
torch==2.1.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
scikit-learn==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
